requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
pdfplumber
python-dotenv

